Refactor LLM to use worker
