+++ gllm/model_runner.py
@@ -65,45 +65,3 @@ class ModelRunner():
 
     def free_kv_cache(self, seq: Sequence):
         self.memory_manager.free(seq)
-
-    def stream_inference(self, seq: Sequence):
-        # -------prefill------
-        prefill_start = time.time()
-        if isinstance(self.memory_manager, PrefixMemoryManager):
-            self.memory_manager.pre_allocate_computed_page([seq])
-        seq.to_compute_token_num = len(seq) - seq.computed_token_num
-        self.memory_manager.pre_allocate_page([seq])
-        next_token = self.step_once(InputData([seq], self.memory_manager))[0]
-        seq.append(next_token)
-        seq.computed_token_num += seq.to_compute_token_num
-        seq.to_compute_token_num = 1
-        prefill_end = time.time()
-        # ----prefill end-----
-
-        # ------decode-------
-        decode_start = time.time()
-        while True:
-            print(seq.detokenize_inc(self.tokenizer), end='', flush=True)
-            if seq.is_finish:
-                break
-            self.memory_manager.pre_allocate_page([seq])
-            next_token = self.step_once(
-                InputData([seq], self.memory_manager))[0]
-            seq.computed_token_num += seq.to_compute_token_num
-            seq.append(next_token)
-        print("\n")
-        decode_end = time.time()
-        # ------decode end-------
-        # ------metric---------
-        elapsed_decode_time = decode_end - decode_start
-        elapsed_prefill_time = prefill_end - prefill_start
-        elapsed_time = round(elapsed_decode_time+elapsed_prefill_time, 2)
-        prefill_rate = round(
-            seq.prompt_len / elapsed_prefill_time, 2)
-        decode_rate = round(
-            len(seq[seq.prompt_len:]) / elapsed_decode_time, 2)
-        print(
-            f'#input: {seq.prompt_len} #output: {len(seq[seq.prompt_len:])} '
-            f'elapsed time: {elapsed_time} s rate(prefill/decode): {prefill_rate}/{decode_rate} toks/s\n')
-        # -----metric end--------
-        return self.decode(seq[seq.prompt_len:])
